temporal-io/my-python/MY/pyproject.toml
[project]
name = "temporal-examples"
version = "0.1.0"
description = "Temporal workflow examples from basic patterns to AI integration"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.18.2",
    "tabulate>=0.9.0",
    "python-dotenv>=1.2.1",
    "openai>=2.6.1",
    "openai-agents>=0.4.2",
]
